fix(gio-form-tags-input): call autocomplete options on focus
The autocomplete options will only be called when the component has the
focus. It prevents calling the async options for nothing, like when the
user doesn't interact with the component.

I didn't manage to prevent calling autocomplete options after selecting
a value. In that case, the input loses the focus and re-gains it just
after, triggering the call of autocomplete options